Reorganized File structure, made tests work
